# splitter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdf-dispatch",
    "version": "1.26.0",
    "description": "Self-hosted PDF splitting service. Watches a folder, splits multi-page\nPDFs on barcode/QR code detection, and routes output documents by trigger code.\n\nConfigurable entirely from the web interface; every setting is also accessible\nvia this REST API, making pdf-dispatch scriptable from Home Assistant, n8n,\nNode-RED, shell scripts, and any HTTP client.\n\n## Authentication\n\nAll endpoints except `/healthz` support two methods (checked in order):\n\n1. **API key** *(recommended for scripts)*: `X-API-Key: <key>` header.  \n   The key is auto-generated at startup and shown in the *API access* section\n   of the web interface. Override with the `API_KEY` environment variable.\n\n2. **HTTP Basic auth**: active only when both `APP_USERNAME` and\n   `APP_PASSWORD` environment variables are set.\n\nIf neither is configured the API is open (trusted network assumption).\nIf `X-API-Key` is present but wrong the request is **always rejected**\n— there is no fall-through to Basic auth.\n",
    "contact": {
      "name": "pdf-dispatch on GitHub",
      "url": "https://github.com/Lheriss/pdf-dispatch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This pdf-dispatch instance"
    }
  ],
...
  "security": [
    {
      "apiKey": []
    },
    {
      "basicAuth": []
    },
    {}
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Auto-generated 64-character hex key. Visible (and regeneratable) in the *API access* section of the web interface. Override with the `API_KEY` environment variable.\n"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Active only when both `APP_USERNAME` and `APP_PASSWORD` environment variables are set.\n"
      }
    },
    "schemas": {
      "OkTrue": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "description": "Human-readable error description (in the active UI language)."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/email/configs": {
      "post": {
        "summary": "Create IMAP configuration",
        "description": "Adds a new email polling configuration.",
        "operationId": "createEmailConfig",
        "tags": [
          "email"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "host",
                  "port",
                  "username"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "host": {
                    "type": "string"
                  },
                  "port": {
                    "type": "integer",
                    "default": 993
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "description": "Stored encrypted."
                  },
                  "folder": {
                    "type": "string",
                    "default": "INBOX"
                  },
                  "verify_ssl": {
                    "type": "boolean",
                    "default": true
                  },
                  "enabled": {
                    "type": "boolean",
                    "default": true
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "read",
                      "delete",
                      "ignore"
                    ],
                    "default": "read",
                    "description": "Action on messages after download."
                  },
                  "poll_interval": {
                    "type": "integer",
                    "default": 5,
                    "description": "Minutes between polls."
                  },
                  "filter_from": {
                    "type": "string",
                    "description": "Filter by sender (contains)."
                  },
                  "filter_subject": {
                    "type": "string",
                    "description": "Filter by subject (contains)."
                  },
                  "default_trigger": {
                    "type": "string",
                    "description": "Trigger for barcode-free attachments."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created configuration (password never returned).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "config": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate name or identical configuration already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/email/configs/{config_id}": {
      "post": {
        "summary": "Update IMAP configuration",
        "operationId": "updateEmailConfig",
        "tags": [
          "email"
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Fields to update (same schema as create; omit unchanged fields)."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "config": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate name or configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete IMAP configuration",
        "operationId": "deleteEmailConfig",
        "tags": [
          "email"
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkTrue"
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/email/test": {
      "post": {
        "summary": "Test IMAP connection",
        "description": "Connects to the IMAP server and counts messages in the specified folder. The connection is not saved. Supply either `password` (clear text, for new configurations) or `id` (reuse the stored encrypted password).\n",
        "operationId": "testEmailConnection",
        "tags": [
          "email"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "host",
                  "port",
                  "username"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "host": {
                    "type": "string"
                  },
                  "port": {
                    "type": "integer"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "id": {
                    "type": "string",
                    "description": "Config ID to reuse stored password."
                  },
                  "folder": {
                    "type": "string",
                    "default": "INBOX"
                  },
                  "verify_ssl": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection successful.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "example": "3 message(s) in INBOX"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Connection failed (wrong credentials, network error, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/email/reset_ids/{config_id}": {
      "post": {
        "summary": "Reset processed message IDs",
        "description": "Clears the processed Message-ID list and lifts the `polling_blocked`\nflag if set. Messages already processed may be re-downloaded on the\nnext poll cycle depending on the configured action.\n\nUse after adjusting an over-broad filter or after a long inactivity period.\n",
        "operationId": "resetEmailIds",
        "tags": [
          "email"
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IDs reset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkTrue"
                }
              }
            }
          },
          "404": {
            "description": "Configuration not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}